Instance controller package name refactor
