+++ operators/pkg/instctrl/exposition.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package instance_controller
+package instctrl
 
 import (
 	"context"
