+++ operators/pkg/instctrl/statusinspection.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package instance_controller
+package instctrl
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
